Crate amethyst_window

Crate abstracting and seperating out the window and display handling within amethyst, and as such its usage of winit.

Structs

DisplayConfig
Configuration for a window display.

EventsLoopSystem
System that polls the window events and pushes them to appropriate event channels.

Icon
An icon used for the window titlebar, taskbar, etc.

MonitorIdent
Identifier for a given monitor. Because there is no cross platform method to actually uniquely identify monitors, this tuple wraps two identifiers of a monitor which should prove sufficient on any given system: The index of the monitor is the retrieved Display array, and the name of the given monitor.

ScreenDimensions
World resource that stores screen dimensions.

Window
Represents a window.

WindowBundle
Bundle providing easy initializing of the appopriate Window, WindowSystem EventLoop and EventLoopSystem constructs used for creating the rendering window of amethyst with winit

WindowSystem
System for opening and managing the window.

Traits

MonitorsAccess
A struct that can resolve monitors. Usually either a Window or an EventsLoop.