Struct amethyst_window::Window

pub struct Window { /* fields omitted */ }

Represents a window.

Example

use winit::{Event, EventsLoop, Window, WindowEvent, ControlFlow};

let mut events_loop = EventsLoop::new();
let window = Window::new(&events_loop).unwrap();

events_loop.run_forever(|event| {
    match event {
        Event::WindowEvent { event: WindowEvent::CloseRequested, .. } => {
            ControlFlow::Break
        },
        _ => ControlFlow::Continue,
    }
});

Methods

impl Window

pub fn new(events_loop: &EventsLoop) -> Result<Window, CreationError>

Creates a new Window for platforms where this is appropriate.

This function is equivalent to WindowBuilder::new().build(events_loop).

Error should be very rare and only occur in case of permission denied, incompatible system, out of memory, etc.

pub fn set_title(&self, title: &str)

Modifies the title of the window.

This is a no-op if the window has already been closed.

pub fn show(&self)

Shows the window if it was hidden.

Platform-specific

Has no effect on Android

pub fn hide(&self)

Hides the window if it was visible.

Platform-specific

Has no effect on Android

pub fn get_position(&self) -> Option<LogicalPosition>

Returns the position of the top-left hand corner of the window relative to the top-left hand corner of the desktop.

Note that the top-left hand corner of the desktop is not necessarily the same as the screen. If the user uses a desktop with multiple monitors, the top-left hand corner of the desktop is the top-left hand corner of the monitor at the top-left of the desktop.

The coordinates can be negative if the top-left hand corner of the window is outside of the visible screen region.

Returns None if the window no longer exists.

pub fn get_inner_position(&self) -> Option<LogicalPosition>

Returns the position of the top-left hand corner of the window's client area relative to the top-left hand corner of the desktop.

The same conditions that apply to get_position apply to this method.

pub fn set_position(&self, position: LogicalPosition)

Modifies the position of the window.

See get_position for more information about the coordinates.

This is a no-op if the window has already been closed.

pub fn get_inner_size(&self) -> Option<LogicalSize>

Returns the logical size of the window's client area.

The client area is the content of the window, excluding the title bar and borders.

Converting the returned LogicalSize to PhysicalSize produces the size your framebuffer should be.

Returns None if the window no longer exists.

pub fn get_outer_size(&self) -> Option<LogicalSize>

Returns the logical size of the entire window.

These dimensions include the title bar and borders. If you don't want that (and you usually don't), use get_inner_size instead.

Returns None if the window no longer exists.

pub fn set_inner_size(&self, size: LogicalSize)

Modifies the inner size of the window.

See get_inner_size for more information about the values.

This is a no-op if the window has already been closed.

pub fn set_min_dimensions(&self, dimensions: Option<LogicalSize>)

Sets a minimum dimension size for the window.

pub fn set_max_dimensions(&self, dimensions: Option<LogicalSize>)

Sets a maximum dimension size for the window.

pub fn set_resizable(&self, resizable: bool)

Sets whether the window is resizable or not.

Note that making the window unresizable doesn't exempt you from handling Resized, as that event can still be triggered by DPI scaling, entering fullscreen mode, etc.

Platform-specific

This only has an effect on desktop platforms.

Due to a bug in XFCE, this has no effect on Xfwm.

pub fn get_hidpi_factor(&self) -> f64

Returns the DPI factor that can be used to map logical pixels to physical pixels, and vice versa.

See the dpi module for more information.

Note that this value can change depending on user action (for example if the window is moved to another screen); as such, tracking WindowEvent::HiDpiFactorChanged events is the most robust way to track the DPI you need to use to draw.

Platform-specific

X11: Can be overridden using the WINIT_HIDPI_FACTOR environment variable.

Android: Always returns 1.0.

pub fn set_cursor(&self, cursor: MouseCursor)

Modifies the mouse cursor of the window. Has no effect on Android.

pub fn set_cursor_position(
&self,
position: LogicalPosition
) -> Result<(), String>

Changes the position of the cursor in window coordinates.

pub fn grab_cursor(&self, grab: bool) -> Result<(), String>

Grabs the cursor, preventing it from leaving the window.

Platform-specific

On macOS, this presently merely locks the cursor in a fixed location, which looks visually awkward.

This has no effect on Android or iOS.

pub fn hide_cursor(&self, hide: bool)

Hides the cursor, making it invisible but still usable.

Platform-specific

On Windows and X11, the cursor is only hidden within the confines of the window.

On macOS, the cursor is hidden as long as the window has input focus, even if the cursor is outside of the window.

This has no effect on Android or iOS.

pub fn set_maximized(&self, maximized: bool)

Sets the window to maximized or back

pub fn set_fullscreen(&self, monitor: Option<MonitorId>)

Sets the window to fullscreen or back

pub fn set_decorations(&self, decorations: bool)

Turn window decorations on or off.

pub fn set_always_on_top(&self, always_on_top: bool)

Change whether or not the window will always be on top of other windows.

pub fn set_window_icon(&self, window_icon: Option<Icon>)

Sets the window icon. On Windows and X11, this is typically the small icon in the top-left corner of the titlebar.

For more usage notes, see WindowBuilder::with_window_icon.

Platform-specific

This only has an effect on Windows and X11.

pub fn set_ime_spot(&self, position: LogicalPosition)

Sets location of IME candidate box in client area coordinates relative to the top left.

pub fn get_current_monitor(&self) -> MonitorId

Returns the monitor on which the window currently resides

pub fn get_available_monitors(&self) -> AvailableMonitorsIter

Returns the list of all the monitors available on the system.

This is the same as EventsLoop::get_available_monitors, and is provided for convenience.

pub fn get_primary_monitor(&self) -> MonitorId

Returns the primary monitor of the system.

This is the same as EventsLoop::get_primary_monitor, and is provided for convenience.

pub fn id(&self) -> WindowId