Struct amethyst_window::WindowBundle

pub struct WindowBundle { /* fields omitted */ }

Methods

impl WindowBundle

pub fn from_config(config: DisplayConfig) -> Self

Builds a new window bundle from a loaded DisplayConfig.

pub fn from_config_path(path: impl AsRef<Path>) -> Self

Builds a new window bundle by loading the DisplayConfig from path.

Will fall back to DisplayConfig::default() in case of an error.